backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
openai
reportlab
pillow
python-dotenv
aiofiles